support for fractional zoom levels
Also changes the way OpenGL textures are handled,
from padding to 4 to using arbitrary texture sizes.

This makes the variables TexWidth, TexHeight, TexMaxS
and TexMaxT obsolete in theory (as they will always
be equal to ScreenWith, ScreenHeight, and 1.0), but
let's just keep them for now.


git-svn-id: https://svn.emphy.de/impressive/trunk/impressive@217 58003caa-664d-4c56-a378-6d187ff66ee1

+++ src/control.py
@@ -330,7 +330,7 @@ def EnterZoomMode(targetx, targety):
     gl.BindTexture(gl.TEXTURE_2D, Tcurrent)
     while gl.GetError():
         pass  # clear all OpenGL errors
-    gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGB, ZoomFactor * TexWidth, ZoomFactor * TexHeight, 0, gl.RGB, gl.UNSIGNED_BYTE, PageImage(Pcurrent, True))
+    gl.TexImage2D(gl.TEXTURE_2D, 0, gl.RGB, int(ZoomFactor * TexWidth), int(ZoomFactor * TexHeight), 0, gl.RGB, gl.UNSIGNED_BYTE, PageImage(Pcurrent, True))
     if gl.GetError():
         print >>sys.stderr, "I'm sorry, but your graphics card is not capable of rendering presentations"
         print >>sys.stderr, "in this resolution. Either the texture memory is exhausted, or there is no"
